requirements.txt
paho-mqtt==1.4.0
#
Adafruit_DHT==1.4.0
